PIL 1763 Update glossary (#14)
* Update index.md

* Update index.md

* Update index.md

* Update index.md

* Update index.md

* Update index.md

* Update index.md

* Update index.md

* Update index.md

* Update index.md

* Update index.md

* Update index.md

Line 152 - updated to
The <em>obligationMTT</em> field cannot be set to true for a domestic only group.
Line 203 - amended to "044".
Line 209 - If the amend request is successful, it returns a response containing a processing date and a charge reference (unless the amendment changes a liability return to a Nil Return, where no charge reference is issued).

Changelog header removed (currently empty). 

Developer hub links proofed and updated.

* Update index.md

* Update index.md